Raise the version of `asm`
The binary compatibility check fails for files compiled with the
Java 9 bytecode or newer as the target. The new `asm` doesn't have
this problem. This is part of the effort to support JDK11 because
the main reason to upgrade to it is to allow for compilation with
bytecode targets newer than 9.